+++ xblock/fields.py
@@ -3,8 +3,9 @@ Fields declare storage for XBlock data.  They use abstract notions of
 **scopes** to associate each field with particular sets of blocks and users.
 The hosting runtime application decides what actual storage mechanism to use
 for each scope.
-
 """
+from __future__ import annotations
+
 from collections import namedtuple
 import copy
 import datetime
@@ -16,6 +17,8 @@ import re
 import time
 import traceback
 import warnings
+
+from opaque_keys.edx.keys import UsageKey, DefinitionKey
 
 import dateutil.parser
 from lxml import etree
@@ -249,6 +252,27 @@ class ScopeIds(namedtuple('ScopeIds', 'user_id block_type def_id usage_id')):
     for instance, the `def_id` identifies scopes that use BlockScope.DEFINITION.
     """
     __slots__ = ()
+
+    def validate_types(self):
+        """
+        Raise an AssertionError if any of the ids are an unexpected type.
+
+        Originally, these fields were all freely-typed; but in practice,
+        edx-platform's XBlock runtime would fail if the ids did not match the
+        types below. In order to make the XBlock library reflect the
+        edx-platform reality and improve type-safety, we've decided to actually
+        enforce the types here, per:
+        https://github.com/openedx/XBlock/issues/708
+        """
+        if self.user_id is not None:
+            if not isinstance(self.user_id, (int, str)):
+                raise TypeError(f"got {self.user_id=}; should be an int, str, or None")
+        if not isinstance(self.block_type, str):
+            raise TypeError(f"got {self.block_type=}; should be a str")
+        if not isinstance(self.def_id, DefinitionKey):
+            raise TypeError(f"got {self.def_id=}; should be a DefinitionKey")
+        if not isinstance(self.usage_id, UsageKey):
+            raise TypeError(f"got {self.usage_id=}; should be a UsageKey")
 
 
 # Define special reference that can be used as a field's default in field